Extract & handle AccessException
